Задание2. Шардирование
