+++ mongo-sharding/api_app/requirements.txt
@@ -1,0 +1,6 @@
+fastapi==0.110.2
+uvicorn[standard]==0.29.0
+motor==3.5.3
+redis==4.4.2
+fastapi-cache2==0.2.0
+logmiddleware==0.0.4
